% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_log_OR.R
\name{CI_log_OR}
\alias{CI_log_OR}
\title{confidence interval for log OR}
\usage{
CI_log_OR(log_OR, numerator, denominator, sig = 0.05, upper = TRUE)
}
\arguments{
\item{log_OR}{log odds ratios}

\item{numerator}{the column whose data are the frequencies of the numerator}

\item{denominator}{the column whose data are the frequencies of the denominator}

\item{sig}{significance level (i.e., 0.05 for 95\% CI)}

\item{upper}{logical, if TRUE (the default), return the upper limit of the confidence interval. Otherwise return its lower limit}
}
\value{
A vector of the confidence limit of the log odds ratios (either the upper or lower limit)
}
\description{
\code{CI_log_OR()} calculates confidence intervals for log odds ratios from the log odds ratios
and data of two vectors representing the frequencies of: numerator; denominator.
}
\examples{
CI_log_OR(rnorm(10), rbinom(10, 1000, 0.8), rbinom(10, 1000, 0.1))

}
