% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_OR.R
\name{compute_OR}
\alias{compute_OR}
\title{compute_OR}
\usage{
compute_OR(numerator, denominator, correction = TRUE)
}
\arguments{
\item{numerator}{the column whose data are the frequencies of the numerator}

\item{denominator}{the column whose data are the frequencies of the denominator}

\item{correction}{logical, if TRUE (the default) the odds ratio will be calculated by adding 0.5 to avoid meaningless results}
}
\value{
A vector of odds ratios
}
\description{
\code{compute_OR()} calculates odds ratios from data of two vectors representing the
frequencies of: numerator; denominator. The correction may be used if there is any frequency equals
zero to avoid getting meaningless outputs.
}
\examples{
compute_OR(rbinom(10, 1000, 0.8), rbinom(10, 1000, 0.1), FALSE)

}
