% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lbw}
\alias{lbw}
\title{Low Birth Weight}
\source{
Modified from Hosmer, D and S. Lemeshow (2000), Applied Logistic Regression, Wiley.
}
\usage{
data(lbw)
}
\value{
A data frame with 189 rows and 11 variables.
}
\description{
The is a modified version of the low birth weight (lbw) data, obtained from Hosmer and
Lemeshow (2000).
}
\details{
\describe{
\item{id}{Subject's identifier}
\item{low}{1=low birthweight baby (under 2500g); 0=norml weight (2500g or over)}
\item{age}{age of mother in years}
\item{lwt}{weight of mother (lbs) at last menstrual period}
\item{race}{factor: white; black; 3=other}
\item{smoke}{1=history of mother smoking; 0=mother nonsmoker}
\item{ptl}{number of false of premature labors: 0-3}
\item{ht}{1=history of hypertension; 0 =no hypertension}
\item{ui}{1=uterine irritability; 0 no irritability}
\item{ftv}{number of physician visits in 1st trimester}
\item{bwt}{birth weight in grams}
}
}
\examples{
data(lbw)
}
\keyword{datasets}
