% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nhanes}
\alias{nhanes}
\title{National Health & Nutrition Examination Survey}
\source{
selected variables from the NHANES data in the \code{NHANES} R package. It is
originally assembled by Michelle Dalrymple of Cashmere High School and Chris Wild of
the University of Auckland, New Zealand for use in teaching statistics.
}
\usage{
data(nhanes)
}
\value{
A data frame with 10,000 rows and 20 variables.
}
\description{
Selected variables from the NHANES 2009-2012 data collected by the US National Center for
Health Statistics (NCHS).
}
\details{
\describe{
\item{ID}{Subject's identifier}
\item{Gender}{Gender (sex) of study participant	coded as male or female}
\item{Age}{Age in years at screening of study participant. Note: Subjects 80 years or older were recorded as 80.}
\item{AgeDecade}{Categorical variable derived from age with levels 0-9, 10-19, ... 70+}
\item{Education}{Educational level of study participant Reported for participants aged 20 years or older.}
\item{MaritalStatus}{Marital status of study participant. Reported for participants aged 20 years or older.}
\item{HomeRooms}{How many rooms are in home of study participant (counting kitchen but not bathroom). 13 rooms = 13 or more rooms.}
\item{HomeOwn}{One of Home, Rent, or Other indicating whether the home of study participant or someone in their family is owned, rented or occupied by some other arrangement.}
\item{Weight}{Weight in kg}
\item{Height}{Standing height in cm. Reported for participants aged 2 years or older.}
\item{BMI}{Body mass index. Reported for participants aged 2 years or older.}
\item{BPSysAve}{Combined systolic blood pressure reading.}
\item{BPDiaAve}{Combined diastolic blood pressure reading}
\item{SleepTrouble}{Study participant has sleep trouble}
\item{Diabetes}{Study participant told by a doctor or health professional that they have diabetes. Reported for participants aged 1 year or older.}
\item{DiabetesAge}{Age of study participant when first told they had diabetes.}
\item{Depressed}{Self-reported number of days where participant felt down, depressed or hopeless. Reported for participants aged 18 years or older. One of None, Several, Majority (more than half the days), or AlmostAll.}
\item{SleepHrsNight}{Self-reported number of hours study participant usually gets at night on weekdays or workdays. Reported for participants aged 16 years and older.}
\item{SmokeNow}{Study participant currently smokes cigarettes regularly. Reported for participants aged 20 years or older}
\item{SmokeAge}{Age study participant first started to smoke cigarettes fairly regularly. Reported for participants aged 20 years or older.}
}
}
\examples{
data(nhanes)
}
\keyword{datasets}
