% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PR_Oct_Deaths}
\alias{PR_Oct_Deaths}
\title{PR October Deaths 2015-2017}
\source{
These data were extracted by Osama Mahmoud (the author and the maintainer of this
R package) from a pdf report on a daily mortality in the month of October through 2015-2018.
The pdf report was available as part of the dslabs R package by Rafael A. Irizarry; https://cran.r-project.org/web/packages/dslabs/dslabs.pdf.
}
\usage{
data(PR_Oct_Deaths)
}
\value{
A character vector with 43 entries.
}
\description{
Daily mortality data for Puerto Rico, the USA territory, extracted for the month of October
through a number of years (2015-2018). The PDF file containing the whole data from
Jan 1, 2015 to May 31, 2018 can be found at: \code{system.file("extdata", "PR-Mortality.pdf", package = "dsEssex")}
}
\examples{
data(PR_Oct_Deaths)
}
\keyword{datasets}
