% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Trump_tweets}
\alias{Trump_tweets}
\title{Trump Tweets from 2009 to 2021}
\source{
Extracted from the Trump Twitter Archive: https://www.thetrumparchive.com by Osama
Mahmoud (the author and the maintainer of this R package)
}
\usage{
data(Trump_tweets)
}
\value{
An object of class "data.frame" with 35,136 rows and 9 variables.
}
\description{
All tweets from Donald Trump's twitter account from 2009 to 8 JAN 2021 (when his account
was suspended post the 2020 US presidential election).
}
\details{
\describe{
\item{id}{Tweet ID}
\item{text}{Tweet}
\item{isRetweet}{logical: whether the text is a retweet}
\item{isDeleted}{logical: whether the text was deleted}
\item{device}{The device/platform from which the tweet was composed}
\item{favorites}{Number of times the tweet had been favored until 8 JAN 2021}
\item{retweets}{Number of times the tweet had been retweeted until 8 JAN 2021}
\item{isFlagged}{logical: whether the tweet was flagged}
\item{created_at}{Date and time at which the tweet was composed}
}
}
\examples{
data(Trump_tweets)
head(Trump_tweets)
}
\keyword{datasets}
