% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tweets_david}
\alias{tweets_david}
\title{Twitter Archive for David Robinson}
\source{
https://github.com/dgrtwo/tidy-text-mining/tree/master/data
}
\usage{
data(tweets_david)
}
\description{
This is the Twitter archive of David Robinson (https://twitter.com/drob) which contains the entire
archive of his tweets.
}
\examples{
data(tweets_david)
}
