% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tweets_julia}
\alias{tweets_julia}
\title{Twitter Archive for Julia Slige}
\source{
https://github.com/dgrtwo/tidy-text-mining/tree/master/data
}
\usage{
data(tweets_julia)
}
\description{
This is the Twitter archive of Julia Silge (https://twitter.com/juliasilge) which contains the entire
archive of her tweets.
}
\examples{
data(tweets_julia)
}
